+++ test/contracts/ERC721GenTest.sol
@@ -4,7 +4,7 @@ import "../../contracts/ERC721Gen.sol";
 
 contract ERC721GenTest {
     event TokenId(uint value);
-    event Traits(uint16[] traits);
+    event Traits(uint[] traits);
 
     function mintAndGetTraits(ERC721Gen token, address to, bool getTraits) public {
         uint tokenId = token.mint(to);
